+++ src/resources/users/user.service.ts
@@ -4,7 +4,9 @@ import  usersRepo from './user.memory.repository';
 
 const getAll = ():Promise<User[]> => usersRepo.getAll();
 
-const createUser = (user: User):Promise<User> => usersRepo.createUser(user);
+const createUser = (user: User):Promise<User | number> => usersRepo.createUser(user);
+
+const createAdmin = (user: User):Promise<User | number> => usersRepo.createAdmin(user);
 
 const getById = (id: number):Promise<User|undefined> => usersRepo.getById(id);
 
@@ -12,5 +14,5 @@ const putById = (newUser: User, id: number):Promise<User|undefined> => usersRepo
 
 const deleteById = (id: number): Promise<number> => usersRepo.deleteById(id);
 
-export default { getAll, createUser, getById, putById, deleteById };
+export default { getAll, createUser, getById, putById, deleteById,createAdmin };
